+++ xknx/knxip/__init__.py
@@ -35,6 +35,7 @@ from .knxip_enum import (
 )
 from .routing_busy import RoutingBusy
 from .routing_indication import RoutingIndication
+from .routing_lost_message import RoutingLostMessage
 from .search_request import SearchRequest
 from .search_request_extended import SearchRequestExtended
 from .search_response import SearchResponse
@@ -80,6 +81,7 @@ __all__ = [
     "KNXMedium",
     "RoutingBusy",
     "RoutingIndication",
+    "RoutingLostMessage",
     "SearchRequest",
     "SearchRequestExtended",
     "SearchRequestParameterType",